+++ Source/PolygonalMapGenerator/Public/Moisture/IslandMoisture.h
@@ -38,18 +38,18 @@ protected:
 	virtual TSet<FPointIndex> FindLakeshores(UTriangleDualMesh* Mesh, const TArray<bool>& r_ocean, const TArray<bool>& r_water) const;
 
 	virtual TSet<FPointIndex> FindMoistureSeeds_Implementation(UTriangleDualMesh* Mesh, const TArray<int32>& s_flow, const TArray<bool>& r_ocean, const TArray<bool>& r_water) const;
-	virtual void AssignRegionMoisture_Implementation(TArray<int32>& r_moisture, TArray<int32>& r_waterdistance, UTriangleDualMesh* Mesh, const TArray<bool>& r_water, const TSet<FPointIndex>& r_moisture_seeds) const;
-	virtual void RedistributeRegionMoisture_Implementation(TArray<int32>& r_moisture, UTriangleDualMesh* Mesh, const TArray<bool>& r_water, float MinMoisture, float MaxMoisture) const;
+	virtual void AssignRegionMoisture_Implementation(TArray<float>& r_moisture, TArray<int32>& r_waterdistance, UTriangleDualMesh* Mesh, const TArray<bool>& r_water, const TSet<FPointIndex>& r_moisture_seeds) const;
+	virtual void RedistributeRegionMoisture_Implementation(TArray<float>& r_moisture, UTriangleDualMesh* Mesh, const TArray<bool>& r_water, float MinMoisture, float MaxMoisture) const;
 
 public:
 	UFUNCTION(BlueprintCallable, BlueprintNativeEvent, Category = "Procedural Generation|Island Generation|Moisture")
 	TSet<FPointIndex> FindMoistureSeeds(UTriangleDualMesh* Mesh, const TArray<int32>& SideFlow, const TArray<bool>& OceanRegions, const TArray<bool>& WaterRegions) const;
 	UFUNCTION(BlueprintCallable, BlueprintNativeEvent, Category = "Procedural Generation|Island Generation|Moisture")
-	void AssignRegionMoisture(UPARAM(ref) TArray<int32>& RegionMoisture, UPARAM(ref) TArray<int32>& RegionWaterDistance, UTriangleDualMesh* Mesh, const TArray<bool>& WaterRegions, const TSet<FPointIndex>& MoistureSeedRegions) const;
+	void AssignRegionMoisture(UPARAM(ref) TArray<float>& RegionMoisture, UPARAM(ref) TArray<int32>& RegionWaterDistance, UTriangleDualMesh* Mesh, const TArray<bool>& WaterRegions, const TSet<FPointIndex>& MoistureSeedRegions) const;
 	UFUNCTION(BlueprintCallable, BlueprintNativeEvent, Category = "Procedural Generation|Island Generation|Moisture")
-	void RedistributeRegionMoisture(UPARAM(ref) TArray<int32>& RegionMoisture, UTriangleDualMesh* Mesh, const TArray<bool>& WaterRegions, float MinMoisture, float MaxMoisture) const;
+	void RedistributeRegionMoisture(UPARAM(ref) TArray<float>& RegionMoisture, UTriangleDualMesh* Mesh, const TArray<bool>& WaterRegions, float MinMoisture, float MaxMoisture) const;
 	
 	TSet<FPointIndex> find_moisture_seeds_r(UTriangleDualMesh* Mesh, const TArray<int32>& s_flow, const TArray<bool>& r_ocean, const TArray<bool>& r_water) const;
-	void assign_r_moisture(TArray<int32>& r_moisture, TArray<int32>& r_waterdistance, UTriangleDualMesh* Mesh, const TArray<bool>& r_water, const TSet<FPointIndex>& r_moisture_seeds) const;
-	void redistribute_r_moisture(TArray<int32>& r_moisture, UTriangleDualMesh* Mesh, const TArray<bool>& r_water, float MinMoisture, float MaxMoisture) const;
+	void assign_r_moisture(TArray<float>& r_moisture, TArray<int32>& r_waterdistance, UTriangleDualMesh* Mesh, const TArray<bool>& r_water, const TSet<FPointIndex>& r_moisture_seeds) const;
+	void redistribute_r_moisture(TArray<float>& r_moisture, UTriangleDualMesh* Mesh, const TArray<bool>& r_water, float MinMoisture, float MaxMoisture) const;
 };
